password_hash2
Fixed the User class's constructor, now it can determine whether the password is already hashed or not. fromDatabase function is also used now, it is the one that will 're-build' user objects from the database, and it will override the default boolean false parameter of the constructor, so rehashing will no longer occur.