password-hash3
Fixed the password_hash of User class's password property. Now when we want to input a new User, before saving it to database we must first create a User object and then use the save() function. The password will be then hashed properly. Now password_hash works for both User creation and on the actual Login Page.