ADMIN CREATE TICKET NOW USES LICENSE_NUMBER
On AdminCreateTicket.php, it will now require an input of the license_number instead of license_id. createTicket() of ticketAPI.php now will check if license_number is the one passed via POST instead of license_id, and if so, it will find the license_id in licenses table using the passed license_number. This way, the save() function of TicketViolations.php class will remain unchanged, as well as script.js.