CRUD License done
Admin CRUD for licenses is done. license number is read only during UPDATE, the rest of the information can be updated.
ERROR: At Admin CRUD for Vehicles, the plate number cannot be updated for some reason
MISSING: still no functions for the default_password, and no admin pages for it